Display only the words that the user failed on
userHasCleared() is now a variable with the same name because it is no longer
possible to determine whether the user has cleared current question. This is
because checkMatchBooleanArray() resets this.synonyms and this.cleared_synonyms.

getRandomArbitrary() and GetJSONFromServer() has been moved to a separate file

some functions in AnswerDescriptionPair have been made private as they are
currently not needed outside

New functions:
isSynonymEqual(x:Synonym , y:Synonym)

getUnclearedSynonyms()